Adds a section to README
Adds a Doing More section with a couple of references where I know that I've shared custom functions here. I'm not sure this is the best place to put this. 

I think that many issues might be helped by people having more familiarity with custom functions and there use.  Is there a "it worked for me" or gallery section where users can post successful function examples with their output and the problem they were solving? I see this on SO, but maybe users are going here first?